requirements.txt
click==8.1.7
Flask>=2.2.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==2.0.1
pymongo==4.6.3
Werkzeug==3.1.3
filetype~=1.2.0
validators~=0.34.0
undetected-chromedriver~=3.5.5
talkingheads~=0.5.4
google-generativeai==0.8.3
google~=3.0.0
protobuf~=5.28.3
pillow~=10.4.0
requests~=2.32.0
nltk~=3.9.1
numpy~=2.2.0
beautifulsoup4~=4.12.3
matplotlib~=3.9.1
Flask-Cors~=3.0.10